+++ src/main/java/com/example/travelbag/domain/restaurant/entity/Restaurant.java
@@ -1,4 +1,4 @@
-package com.example.travelbag.domain.Souvenir.entity;
+package com.example.travelbag.domain.restaurant.entity;
 
 import com.example.travelbag.domain.location.entity.Location;
 import jakarta.persistence.*;
@@ -7,19 +7,20 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-
 @Entity
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class Souvenir {
+public class Restaurant {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
 
     private String name;
+
+    private String signature;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "location_id")
