+++ src/main/java/com/example/travelbag/domain/souvenir/controller/api/SouvenirApi.java
@@ -1,6 +1,6 @@
-package com.example.travelbag.domain.Restaurant.controller.api;
+package com.example.travelbag.domain.souvenir.controller.api;
 
-import com.example.travelbag.domain.Restaurant.dto.RestaurantResponseDTO;
+import com.example.travelbag.domain.souvenir.dto.SouvenirResponseDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
@@ -12,18 +12,18 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.List;
 
-@Tag(name = "Restaurant", description = "맛집 관련 API")
-@RequestMapping("/restaurant")
-public interface RestaurantApi {
+@Tag(name = "Souvenir", description = "기념품 관련 API")
+@RequestMapping("/souvenir")
+public interface SouvenirApi {
 
     @Operation(
-            summary = "여행지별 추천 맛집 조회",    // 짧은 설명
-            description = "여행지별 추천 맛집 목록을 조회합니다.",    // 상세 설명
+            summary = "여행지별 추천 기념품 조회",    // 짧은 설명
+            description = "여행지별 추천 기념품 목록을 조회합니다.",    // 상세 설명
             security = @SecurityRequirement(name = "bearerAuth")   // API 호출 시 인증 요구사항(추후 수정 예정)
     )
     // 반환 상태 코드 및 의미
-    @ApiResponse(responseCode = "200", description = "맛집 목록 조회 성공")
+    @ApiResponse(responseCode = "200", description = "기념품 목록 조회 성공")
     @ApiResponse(responseCode = "401", description = "인증 실패")
     @GetMapping()
-    ResponseEntity<List<RestaurantResponseDTO>> getRestaurantsByLocation(@RequestParam(value="location_id") Long location_id);
+    ResponseEntity<List<SouvenirResponseDTO>> getSouvenirsByLocation(@RequestParam(value="location_id") Long location_id);
 }
