+++ src/main/java/com/example/travelbag/domain/souvenir/entity/Souvenir.java
@@ -1,4 +1,4 @@
-package com.example.travelbag.domain.Restaurant.entity;
+package com.example.travelbag.domain.souvenir.entity;
 
 import com.example.travelbag.domain.location.entity.Location;
 import jakarta.persistence.*;
@@ -7,20 +7,19 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+
 @Entity
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class Restaurant {
+public class Souvenir {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
 
     private String name;
-
-    private String signature;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "location_id")
